ft: add `getDereferencedEntityTypesActivity` as activity
